helios-control-room/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.7.1
openpyxl==3.1.5
xlsxwriter==3.2.9
scikit-learn==1.4.2
